move models and service to new core project
