+++ jobs/Backend/Task/ExchangeRateUpdater.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater", "ExchangeRateUpdater.csproj", "{7B2695D6-D24C-4460-A58E-A10F08550CE0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater", "ExchangeRateUpdater\ExchangeRateUpdater.csproj", "{8CB5031E-434E-4E84-9D18-8156EF23F144}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater.Core", "ExchangeRateUpdater.Core\ExchangeRateUpdater.Core.csproj", "{7982F78A-A2CE-40F9-978B-51C00C354C4F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CB5031E-434E-4E84-9D18-8156EF23F144}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CB5031E-434E-4E84-9D18-8156EF23F144}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CB5031E-434E-4E84-9D18-8156EF23F144}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CB5031E-434E-4E84-9D18-8156EF23F144}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7982F78A-A2CE-40F9-978B-51C00C354C4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7982F78A-A2CE-40F9-978B-51C00C354C4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7982F78A-A2CE-40F9-978B-51C00C354C4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7982F78A-A2CE-40F9-978B-51C00C354C4F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
