+++ jobs/Backend/Task/ExchangeRateUpdater/ExchangeRateUpdater.csproj
@@ -5,4 +5,8 @@
     <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ExchangeRateUpdater.Core\ExchangeRateUpdater.Core.csproj" />
+  </ItemGroup>
+
 </Project>
